feat: migrate AI chat from Gemma/Google to Groq (llama-3.3-70b-versatile)
- Replace IGemmaAiService/DirectGemmaService with IGroqAiService/GroqAiService
- Use Groq OpenAI-compatible API with json_object response_format for reliable JSON output
- Fix DietPlanResponse DTO: change int fields to double to handle AI decimal values
- Add (int) casts in AIChatService for database entity assignments
- Reduce conversation history from 10 to 6 messages to reduce token usage
- Update DependencyInjection.cs to register GroqAiService
- Update CreateDietPlanCommandHandler to use IGroqAiService

+++ backend/src/FitnessTrainingSystem.Application/DTOs/Nutrition/DietPlanResponse.cs
@@ -8,16 +8,16 @@ public class DietPlanResponse
     public string DietTitle { get; set; } = "";
 
     [JsonPropertyName("daily_calories")]
-    public int DailyCalories { get; set; }
+    public double DailyCalories { get; set; }
 
     [JsonPropertyName("protein_target_g")]
-    public int ProteinTargetG { get; set; }
+    public double ProteinTargetG { get; set; }
 
     [JsonPropertyName("carbs_target_g")]
-    public int CarbsTargetG { get; set; }
+    public double CarbsTargetG { get; set; }
 
     [JsonPropertyName("fat_target_g")]
-    public int FatTargetG { get; set; }
+    public double FatTargetG { get; set; }
 
     [JsonPropertyName("meals")]
     public List<MealDto> Meals { get; set; } = new();
@@ -29,7 +29,7 @@ public class MealDto
     public string Name { get; set; } = "";
 
     [JsonPropertyName("calories")]
-    public int Calories { get; set; }
+    public double Calories { get; set; }
 
     [JsonPropertyName("foods")]
     public List<FoodItemDto> Foods { get; set; } = new();
@@ -47,7 +47,7 @@ public class FoodItemDto
     public string Amount { get; set; } = "";
 
     [JsonPropertyName("calories")]
-    public int Calories { get; set; }
+    public double Calories { get; set; }
 
     [JsonPropertyName("protein")]
     public double Protein { get; set; }
